+++ family-tree-backend/src/main/java/com/example/tree/repositories/TreeRepository.java
@@ -3,5 +3,5 @@ package com.example.tree.repositories;
 import com.example.tree.model.Tree;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TreeRepository extends CrudRepository<Tree, Integer> {
+public interface TreeRepository extends CrudRepository<Tree, Long> {
 }
